+++ lib/features/main/screens/create_wallet_screen.dart
@@ -1,0 +1,211 @@
+import 'package:flutter/material.dart';
+import 'package:resonance_network_wallet/core/extensions/color_extensions.dart';
+import 'package:resonance_network_wallet/core/services/substrate_service.dart';
+import 'package:resonance_network_wallet/main.dart';
+import 'package:shared_preferences/shared_preferences.dart';
+import 'package:flutter/services.dart';
+
+class CreateWalletScreen extends StatefulWidget {
+  const CreateWalletScreen({super.key});
+
+  @override
+  CreateWalletScreenState createState() => CreateWalletScreenState();
+}
+
+class CreateWalletScreenState extends State<CreateWalletScreen> {
+  String _mnemonic = '';
+  bool _isLoading = false;
+  bool _hasSavedMnemonic = false;
+
+  @override
+  void initState() {
+    super.initState();
+    _generateMnemonic();
+  }
+
+  Future<void> _generateMnemonic() async {
+    setState(() {
+      _isLoading = true;
+    });
+
+    try {
+      _mnemonic = await SubstrateService().generateMnemonic();
+      final walletInfo = await SubstrateService().generateNewWallet(_mnemonic);
+      debugPrint('Generated wallet address: ${walletInfo.accountId}');
+
+      setState(() {
+        _isLoading = false;
+      });
+    } catch (e) {
+      debugPrint('Error generating mnemonic: $e');
+      setState(() {
+        _isLoading = false;
+      });
+    }
+  }
+
+  Future<void> _saveWalletAndContinue() async {
+    try {
+      final walletInfo = await SubstrateService().generateNewWallet(_mnemonic);
+
+      if (mode == Mode.dilithium) {
+        throw Exception('Dilithium is not supported yet');
+      }
+      // Save wallet info
+      final prefs = await SharedPreferences.getInstance();
+      await prefs.setBool('has_wallet', true);
+      await prefs.setString('mnemonic', _mnemonic);
+      await prefs.setString('account_id', walletInfo.accountId);
+
+      if (context.mounted && mounted) {
+        Navigator.pushAndRemoveUntil(
+          context,
+          MaterialPageRoute(builder: (context) => const WalletMain()),
+          (route) => false,
+        );
+      }
+    } catch (e) {
+      debugPrint('Error saving wallet: $e');
+      if (mounted) {
+        ScaffoldMessenger.of(context).showSnackBar(
+          SnackBar(content: Text('Error saving wallet: $e')),
+        );
+      }
+    }
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text('Create New Wallet'),
+      ),
+      body: _isLoading
+          ? const Center(child: CircularProgressIndicator())
+          : Column(
+              children: [
+                Expanded(
+                  child: SingleChildScrollView(
+                    padding: const EdgeInsets.all(16.0),
+                    child: Column(
+                      crossAxisAlignment: CrossAxisAlignment.start,
+                      children: [
+                        const Text(
+                          'Your Recovery Phrase',
+                          style: TextStyle(
+                            fontSize: 24,
+                            fontWeight: FontWeight.bold,
+                          ),
+                        ),
+                        const SizedBox(height: 8),
+                        const Text(
+                          'Write down these words in order and keep them in a safe place. This is the only way to recover your wallet if you lose access.',
+                          style: TextStyle(
+                            color: Colors.red,
+                            fontSize: 14,
+                          ),
+                        ),
+                        const SizedBox(height: 24),
+                        Container(
+                          width: double.infinity,
+                          padding: const EdgeInsets.all(16),
+                          decoration: BoxDecoration(
+                            color: Theme.of(context).cardColor,
+                            borderRadius: BorderRadius.circular(12),
+                            border: Border.all(
+                              color: const Color(0xFF6B46C1).useOpacity(0.3),
+                            ),
+                          ),
+                          child: Column(
+                            crossAxisAlignment: CrossAxisAlignment.start,
+                            children: [
+                              Text(
+                                _mnemonic,
+                                style: const TextStyle(
+                                  fontSize: 18,
+                                  height: 1.5,
+                                  letterSpacing: 0.5,
+                                ),
+                              ),
+                              const SizedBox(height: 16),
+                              Center(
+                                child: OutlinedButton.icon(
+                                  onPressed: () {
+                                    Clipboard.setData(ClipboardData(text: _mnemonic));
+                                    ScaffoldMessenger.of(context).showSnackBar(
+                                      const SnackBar(content: Text('Recovery phrase copied to clipboard')),
+                                    );
+                                  },
+                                  icon: const Icon(Icons.copy),
+                                  label: const Text('Copy to Clipboard'),
+                                ),
+                              ),
+                            ],
+                          ),
+                        ),
+                        const SizedBox(height: 24),
+                        Container(
+                          padding: const EdgeInsets.all(16),
+                          decoration: BoxDecoration(
+                            color: Colors.red.useOpacity(0.1),
+                            borderRadius: BorderRadius.circular(12),
+                          ),
+                          child: const Row(
+                            children: [
+                              Icon(Icons.warning_amber_rounded, color: Colors.red),
+                              SizedBox(width: 12),
+                              Expanded(
+                                child: Text(
+                                  'Never share your recovery phrase with anyone. Store it securely offline.',
+                                  style: TextStyle(color: Colors.red),
+                                ),
+                              ),
+                            ],
+                          ),
+                        ),
+                      ],
+                    ),
+                  ),
+                ),
+                SafeArea(
+                  child: Padding(
+                    padding: const EdgeInsets.all(16),
+                    child: Column(
+                      mainAxisSize: MainAxisSize.min,
+                      children: [
+                        Row(
+                          children: [
+                            Checkbox(
+                              value: _hasSavedMnemonic,
+                              onChanged: (value) {
+                                setState(() {
+                                  _hasSavedMnemonic = value ?? false;
+                                });
+                              },
+                            ),
+                            const Expanded(
+                              child: Text(
+                                'I have written down my recovery phrase and stored it securely',
+                                style: TextStyle(fontWeight: FontWeight.bold),
+                              ),
+                            ),
+                          ],
+                        ),
+                        const SizedBox(height: 16),
+                        SizedBox(
+                          width: double.infinity,
+                          height: 50,
+                          child: ElevatedButton(
+                            onPressed: _hasSavedMnemonic ? _saveWalletAndContinue : null,
+                            child: const Text('Continue'),
+                          ),
+                        ),
+                      ],
+                    ),
+                  ),
+                ),
+              ],
+            ),
+    );
+  }
+}
